feat: User Profile (performance score 50 => 60 (lighthouse)) + my offering (not in postID)(performace score 64 => 73 (lighthouse)) + Optimize Navbar (I don't see much of a performance change on my page, so I'm not sure if this is working or not)
