+++ frontend/src/app/(logged-in)/create-offer/[pid]/page.tsx
@@ -132,7 +132,7 @@ import {useRouter } from "next/navigation";
             // responseOffer= await getOfferHistory(token,userID); 
           }
           setUserData(userResponse)
-          console.log("respons",response)
+          // console.log("respons",response)
           if (response) {
             const posts =
               userRole === "producer" ? response : [response];
