+++ frontend/src/app/layout.tsx
@@ -6,17 +6,21 @@ import NavBar from "@/components/NavBar";
 import Footer from "@/components/Footer";
 import { SessionProvider } from "next-auth/react";
 import { auth } from "@/auth";
-import AuthContext from "@/components/AuthProvider";
 import ReduxProvider from "@/components/ReduxProvider";
+import dynamic from "next/dynamic";
+import Head from "next/head";
+import { useMemo } from "react";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
+  display : 'swap'
 });
 
 const geistMono = Geist_Mono({
   variable: "--font-geist-mono",
   subsets: ["latin"],
+  display : 'swap'
 });
 
 export const metadata: Metadata = {
@@ -24,6 +28,8 @@ export const metadata: Metadata = {
   description: "A Professional Network for Creative Talent",
 };
 
+const NavBar =  dynamic(() => import("@/components/NavBar"));
+const Footer = dynamic(() => import("@/components/Footer"));
 export default async function RootLayout({
   children,
 }: Readonly<{
@@ -33,6 +39,7 @@ export default async function RootLayout({
 
   return (
     <html lang="en">
+      
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
