+++ frontend/src/redux/store.ts
@@ -37,6 +37,6 @@ export const store = configureStore({
     
 })
 
-export type RootState = ReturnType<typeof store.getState>
+export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch
 export const persistor = persistStore(store)
